requirements.txt
coincurve>=15.0.1,<18.0.0
bip-utils>=2.7.0
eth-account>=0.8.0
eth-utils>=2.0.0
tqdm>=4.65.0
PyYAML==6.0.1
web3>=6.0.0
cryptography>=41.0.0
rich>=13.0.0
click>=8.0.0
mnemonic>=0.20
hdwallet>=2.2.1
bip-utils>=2.7.0